fix(todos): fix in todos and make simple pagination

+++ src/modules/todos/repos/todo.repo.ts
@@ -3,6 +3,9 @@ import { Todo } from '../entities/todo.entity';
 import { Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
 import { ITodoRepository } from '../interfaces';
+import { paginate } from 'src/common/pagination/paginate.helper';
+import { PaginationDto } from 'src/common/pagination/dto/requests/pagination.dto';
+import { PaginatedResponseDto } from 'src/common/pagination/dto';
 
 @Injectable()
 export class TodoRepository implements ITodoRepository {
@@ -23,13 +26,20 @@ export class TodoRepository implements ITodoRepository {
     return this.todoRepository.save(todo);
   }
 
-  async getAllTodos() {
-    return this.todoRepository.find();
+  async getAllTodos(
+    userId: string,
+    paginationDto: PaginationDto,
+  ): Promise<PaginatedResponseDto<Todo>> {
+    // return this.todoRepository.find({
+    //   where: { ownerId: userId },
+    // });
+
+    return paginate(this.todoRepository, { ownerId: userId }, paginationDto);
   }
 
-  async findOneById(id: string) {
+  async findOneById(id: string, userId: string): Promise<Todo> {
     const todo = await this.todoRepository.findOne({
-      where: { id },
+      where: { id, ownerId: userId },
       select: [
         'id',
         'title',
@@ -47,15 +57,18 @@ export class TodoRepository implements ITodoRepository {
     return todo;
   }
 
-  async findAndUpdate(id: string, updateData: Partial<Todo>) {
-    const result = await this.todoRepository.update(id, updateData);
+  async findAndUpdate(id: string, updateData: Partial<Todo>, userId: string) {
+    const result = await this.todoRepository.update(
+      { id, ownerId: userId },
+      updateData,
+    );
     if (result.affected === 0) {
       throw new NotFoundException('No todo match provided id');
     }
   }
 
-  async findAndDelete(id: string) {
-    const result = await this.todoRepository.delete(id);
+  async findAndDelete(id: string, userId: string) {
+    const result = await this.todoRepository.delete({ id, ownerId: userId });
     if (result.affected === 0) {
       throw new NotFoundException('No todo match provided id');
     }
